Admin dashboard: new edit plugin form & refactor new plugin form
