+++ src/VAXPRED/src/app/admin-page-plugin-form/admin-page-plugin-form.component.css
@@ -1,0 +1,28 @@
+.add-plugin-content {
+  padding: 2em;
+}
+
+.add-plugin-input-container {
+  width: 100%;
+}
+
+.add-plugin-label {
+  font-size: large;
+}
+
+.add-tag-container {
+  display: flex;
+  align-items: center;
+}
+
+.add-tag-container input {
+  flex: 1;
+  margin-right: 8px;
+}
+
+.tip-message {
+  margin-top: -25px;
+  margin-bottom: 20px;
+  color: #757575;
+  font-size: 12px;
+}
