+++ src/VAXPRED/src/app/plugin-page/plugin-page.component.ts
