+++ src/VAXPRED/src/app/search-page/plugin-card/plugin-card.component.css
