+++ src/VAXPRED/src/app/search-page/plugin-card/plugin-card.component.html
@@ -9,7 +9,7 @@
 		<img id="plugin-type-icon" [src]="this.tool.type_icon" alt="Plugin Type Icon">
 		<span class="tooltip">{{this.tool.type_description}}</span>
 	</div>
-	
+
 	<div class="m-5" >
 		<img id="plugin-image" [src]="this.tool.logo" [alt]="this.tool.name">
 	</div>
